json update and obsevation added

+++ lib/features/patient/patient_assessments_and_tests/domain/entities/triage_report_and_assessment_entity.dart
@@ -37,7 +37,7 @@ class QuestionResponseBriefEntity {
 }
 
 class ObservationBriefEntity {
-  final int observationValue;
+  final double observationValue;
   final int observationId;
   final int observationValueIdentifier;
   final String bodySite;
